requirements.txt
Flask==2.3.3
web3==6.11.0
Flask-SQLAlchemy==3.0.5
eth-account==0.9.0
gunicorn==21.2.0
python-telegram-bot==20.8
aiohttp==3.9.5
matplotlib==3.8.4
pandas==2.2.2
